requirements.txt
Flask==2.3.3
PyGithub==1.59.1
python-dotenv==1.0.0
requests==2.31.0
